fix: normalize INT variable storage to int64 for base decoder parity
handleVar and decodeMapStringAny stored INT tokens as Go int (from
fastParseInt), while the base decoder stores int64. This caused a type
mismatch for map[string]any consumers that assert int64.

Fix by wrapping with int64() at the two storage points. No performance
impact — int64() is a register-only operation on amd64.

Also add TestNeo_VariableType confirming concrete Go types in:
- handleVar variable storage
- decodeMapStringAny map value storage

+++ neo_decoder.go
@@ -181,7 +181,7 @@ func (dec *NeoDecoder) handleVar() error {
 	case STRING:
 		dec.variables[name] = BytesToString(valTok.Literal)
 	case INT:
-		dec.variables[name] = dec.fastParseInt(valTok.Literal)
+		dec.variables[name] = int64(dec.fastParseInt(valTok.Literal))
 	case FLOAT:
 		f64, _ := strconv.ParseFloat(BytesToString(valTok.Literal), 64)
 		dec.variables[name] = f64
@@ -1015,7 +1015,7 @@ func (dec *NeoDecoder) decodeMapStringAny(rv reflect.Value, hasBracket bool) {
 		case STRING:
 			val = BytesToString(valTok.Literal)
 		case INT:
-			val = dec.fastParseInt(valTok.Literal)
+			val = int64(dec.fastParseInt(valTok.Literal))
 		case FLOAT:
 			f64, _ := strconv.ParseFloat(BytesToString(valTok.Literal), 64)
 			val = f64
